requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
httpx==0.28.1
pytesseract==0.3.13
PyMuPDF==1.25.3
Pillow==11.1.0
pydantic==2.10.5
pydantic-settings==2.7.1
